passed 7 of 8 alarm tests for alarm clock

+++ src/threads/thread.h
@@ -96,7 +96,7 @@ struct thread
     
     struct semaphore sema;
     struct list_elem timer_list_elem;
-    int wakeup_time;
+    int64_t wakeup_time;
 
 #ifdef USERPROG
     /* Owned by userprog/process.c. */
